move rows and column local state merging into io/merged
