+++ apps/web/modules/components/editor/blocks/table/table-block-store.tsx
@@ -137,96 +137,25 @@ export class TableBlockStore {
             skip: pageNumber * PAGE_SIZE,
           };
 
-          const { columns: serverColumns } = await this.api.columns({
-            params,
-            abortController: this.abortController,
-          });
-
-          const { rows: serverRows } = await this.api.rows({
-            params,
-            abortController: this.abortController,
-          });
-
           /**
            * Aggregate columns from local and server columns.
            */
-          const columns = EntityTable.columnsFromActions(
-            this.ActionsStore.actions$.get()[spaceId],
-            serverColumns,
-            selectedType?.entityId
-          );
+          const { columns } = await this.MergedData.columns({
+            params,
+            abortController: this.abortController,
+          });
 
           /**
            * Aggregate data for the rows from local and server entities.
-           *
-           * There are several edge-cases we need to handle in order to correctly merge local changes
-           * with server data in the entity table:
-           * 1. An entity is created locally and is given the selected type
-           * 2. An entity is edited locally and is given the selected type
-           * 3. A type is created locally and an entity is given the new type
-           *
-           * Since the table aggregation code expects triples, we may end up in a situation where
-           * the type for an entity has changed, but the name hasn't. In this case there is no local
-           * version of the name triple, so we need to fetch it along with any other triples the table
-           * needs to render the columnSchema.
            */
-          const changedEntitiesIdsFromAnotherType = pipe(
-            this.ActionsStore.actions$.get()[spaceId],
-            actions => Triple.fromActions(actions, []),
-            triples => Entity.entitiesFromTriples(triples),
-            A.filter(e => e.types.some(t => t.id === selectedType?.entityId)),
-            A.map(t => t.id)
+          const { rows } = await this.MergedData.rows(
+            {
+              params,
+              abortController: this.abortController,
+            },
+            columns,
+            selectedType?.entityId
           );
-
-          // Fetch any entities that exist already remotely that have been changed locally
-          // and have the selected type to make sure we have all of the triples necessary
-          // to represent the entity in the table.
-          //
-          // e.g., We add Type A to Entity A. When we render the Type A table, we need
-          // _all_ of the triples for Entity A, not just the ones that have changed locally.
-          //
-          // This will return null if the entity we're fetching does not exist remotely.
-          // i.e., the entity was created locally and has not been published to the server.
-          const maybeServerEntitiesChangedLocally = await Promise.all(
-            changedEntitiesIdsFromAnotherType.map(id => this.api.fetchEntity(id))
-          );
-
-          const serverEntitiesChangedLocally = maybeServerEntitiesChangedLocally.flatMap(e => (e ? [e] : []));
-
-          const serverEntityTriples = serverRows.flatMap(t => t.triples);
-
-          const entitiesCreatedOrChangedLocally = pipe(
-            this.ActionsStore.actions$.get(),
-            actions => Entity.mergeActionsWithEntities(actions, Entity.entitiesFromTriples(serverEntityTriples)),
-            A.filter(e => e.types.some(t => t.id === selectedType?.entityId))
-          );
-
-          const localEntitiesIds = new Set(entitiesCreatedOrChangedLocally.map(e => e.id));
-          const serverEntitiesChangedLocallyIds = new Set(serverEntitiesChangedLocally.map(e => e.id));
-
-          // Filter out any server rows that have been changed locally
-          const filteredServerRows = serverEntityTriples.filter(
-            sr => !localEntitiesIds.has(sr.entityId) && !serverEntitiesChangedLocallyIds.has(sr.entityId)
-          );
-
-          const entities = Entity.entitiesFromTriples([
-            // These are entities that were created locally and have the selected type
-            ...entitiesCreatedOrChangedLocally.flatMap(e => e.triples),
-
-            // These are entities that have a new type locally and may exist on the server.
-            // We need to fetch all triples associated with this entity in order to correctly
-            // populate the table.
-            ...serverEntitiesChangedLocally.flatMap(e => e.triples),
-
-            // These are entities that have been fetched from the server and have the selected type.
-            // They are deduped from the local changes above.
-            ...filteredServerRows,
-          ]);
-
-          // Make sure we only generate rows for entities that have the selected type
-          const entitiesWithSelectedType = entities.filter(e => e.types.some(t => t.id === selectedType?.entityId));
-
-          const { rows } = EntityTable.fromColumnsAndRows(spaceId, entitiesWithSelectedType, columns);
 
           this.isLoading$.set(false);
