+++ apps/web/modules/entity/entity-table-store/Table.ts
@@ -4,7 +4,7 @@ import { Triple } from '~/modules/triple';
 import { Action, Column, Entity, Row } from '~/modules/types';
 import { DEFAULT_PAGE_SIZE } from '..';
 
-export function fromColumnsAndRows(spaceId: string, entities: Entity[], columns: Column[]) {
+export function fromColumnsAndRows(entities: Entity[], columns: Column[]) {
   /* Finally, we can build our initialRows */
   const aggregatedRows = entities.map(({ triples, id }) => {
     return columns.reduce((acc, column) => {
